moved remove quote
