+++ Makefile
@@ -79,7 +79,8 @@ define SRC 	:=
 		create_ast.c \
 		expand_utils.c \
 		expand_var.c \
-		parse_line.c
+		parse_line.c \
+		remove_quotes.c
 	)
 	env.c
 	main.c
