+++ src/parsing/add_dir.c
@@ -6,7 +6,7 @@
 /*   By: abasdere <abasdere@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/25 11:18:39 by abasdere          #+#    #+#             */
-/*   Updated: 2024/02/15 10:40:10 by abasdere         ###   ########.fr       */
+/*   Updated: 2024/02/15 13:16:45 by abasdere         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,6 +29,8 @@ static t_code	add_in(t_ast *ast, t_intype type)
 	if (!file)
 		return (free_args(ast), error(C_MEM, "ft_calloc", M_MEM));
 	file->intype = type;
+	if (type != IT_HERE_DOC)
+		remove_quotes(ast->args[ast->i]);
 	file->filename = ast->args[ast->i];
 	if (addback_cmd(ast->target, new_element(T_INFILE, file)) && ast->i++)
 		return (free_args(ast), free(file->filename),
@@ -52,7 +54,7 @@ static t_code	add_out(t_ast *ast, t_outtype type)
 	if (!file)
 		return (free_args(ast), error(C_MEM, "ft_calloc", M_MEM));
 	file->outtype = type;
-	file->filename = ast->args[ast->i];
+	(remove_quotes(ast->args[ast->i]), file->filename = ast->args[ast->i]);
 	if (addback_cmd(ast->target, new_element(T_INFILE, file)) && ast->i++)
 		return (free_args(ast), free(file->filename),
 			free(file), error(C_MEM, "addback_cmd", M_MEM));
